--- library/traduco/version.rb
@@ -1,3 +1,0 @@
-module Traduco
-  VERSION = '0.9.0'
-end
